remove everything extra
remove everything that will be added to avatar toolkit
