+++ tools/tools.py
@@ -9,598 +9,13 @@ from .. import globals
 
 from .dictionaries import bone_names
 from . import core
-from .translate import t, translation_enum, translate_mmd
+from .translate import t
 
 from ..class_register import wrapper_registry
 
 from bpy.types import Context, Operator
 
 from mathutils.geometry import intersect_point_line
-
-
-@wrapper_registry
-class Tuxedo_OT_DuplicateBones(bpy.types.Operator):
-    bl_idname = 'tuxedo.duplicate_bones'
-    bl_label = t('Tools.duplicate_bones.label')
-    bl_description = t('Tools.duplicate_bones.desc')
-    bl_options = {'REGISTER', 'UNDO'}
-
-    @classmethod
-    def poll(cls, context: bpy.types.Context):
-        return (context.active_object is not None) and (context.object.mode == 'EDIT') and (context.active_object.type == "ARMATURE")
-
-    def execute(self, context: bpy.types.Context):
-        core.dup_and_split_weights_bones(context, context.object)
-        return {'FINISHED'}
-
-@wrapper_registry
-class Tuxedo_OT_ConnectBones(bpy.types.Operator):
-    bl_idname = 'tuxedo.connect_bones'
-    bl_label = t('Tools.connect_bones.label')
-    bl_description = t('Tools.connect_bones.desc')
-    bl_options = {'REGISTER', 'UNDO'}
-
-    @classmethod
-    def poll(cls, context: bpy.types.Context):
-        return (context.active_object is not None) and (context.object.mode == 'OBJECT' or context.object.mode == 'EDIT') and (context.active_object is not None) and (context.active_object.type == "ARMATURE")
-
-    def execute(self, context: bpy.types.Context):
-        core.connect_bones(context, context.object.data)
-        return {'FINISHED'}
-
-@wrapper_registry
-class Tuxedo_OT_RemoveDoublesSafely(bpy.types.Operator):
-    bl_idname = 'tuxedo.remove_doubles_safely'
-    bl_label = t('Tools.remove_doubles_safely.label')
-    bl_description = t('Tools.remove_doubles_safely.desc')
-    bl_options = {'REGISTER', 'UNDO'}
-
-    @classmethod
-    def poll(cls, context: bpy.types.Context):
-        return (context.active_object is not None) and (context.object.mode == 'EDIT') and (context.active_object is not None) and (context.active_object.type == "MESH")
-
-    margin: bpy.props.FloatProperty(default=.00001)
-
-    def execute(self, context: bpy.types.Context):
-        core.remove_doubles_safely(mesh=context.active_object.data, margin=self.margin, merge_all=False)
-
-        return {'FINISHED'}
-    
-@wrapper_registry
-class Tuxedo_OT_CreateDigitigradeLegs(bpy.types.Operator):
-    bl_idname = "armature.createdigitigradelegs"
-    bl_label = t('Tools.create_digitigrade_legs.label')
-    bl_description = t('Tools.create_digitigrade_legs.desc')
-
-    @classmethod
-    def poll(cls, context):
-        if(context.active_object is None):
-            return False
-        if(context.selected_editable_bones is not None):
-            if(len(context.selected_editable_bones) == 2):
-                return True
-        return False
-
-    def execute(self, context):
-
-        for digi0 in context.selected_editable_bones:
-            digi1 = None
-            digi2 = None
-            digi3 = None
-
-            try:
-                digi1 = digi0.children[0]
-                digi2 = digi1.children[0]
-                digi3 = digi2.children[0]
-            except:
-                print("bone format incorrect! Please select a chain of 4 continious bones!")
-            digi4 = None
-            try:
-                digi4 = digi3.children[0]
-            except:
-                print("no toe bone. Continuing.")
-            digi0.select = True
-            digi1.select = True
-            digi2.select = True
-            digi3.select = True
-            if(digi4):
-                digi4.select = True
-            bpy.ops.armature.roll_clear()
-            bpy.ops.armature.select_all(action='DESELECT')
-
-            scene = context.scene
-            #creating transform for upper leg
-            digi0.select = True
-            bpy.ops.transform.create_orientation(name="CATS_digi0", overwrite=True)
-            bpy.ops.armature.select_all(action='DESELECT')
-
-
-            #duplicate digi0 and assign it to thigh
-            thigh = core.duplicatebone(digi0)
-            bpy.ops.armature.select_all(action='DESELECT')
-
-            #make digi2 parrallel to digi1
-            digi2.align_orientation(digi0)
-
-            #extrude thigh
-            thigh.select_tail = True
-            bpy.ops.armature.extrude_move(ARMATURE_OT_extrude={"forked":False},TRANSFORM_OT_translate=None)
-            #set new bone to calf varible
-            bpy.ops.armature.select_more()
-            calf = context.selected_bones[0]
-            bpy.ops.armature.select_all(action='DESELECT')
-
-            #set calf end to  digi2 end
-            calf.tail = digi2.tail
-
-            #make copy of calf, flip it, and then align bone so that it's head is moved to match in align phase
-            flipedcalf = core.duplicatebone(calf)
-            bpy.ops.armature.select_all(action='DESELECT')
-            flipedcalf.select = True
-            bpy.ops.armature.switch_direction()
-            bpy.ops.armature.select_all(action='DESELECT')
-            flippeddigi1 = core.duplicatebone(digi1)
-            bpy.ops.armature.select_all(action='DESELECT')
-            flippeddigi1.select = True
-            bpy.ops.armature.switch_direction()
-            bpy.ops.armature.select_all(action='DESELECT')
-
-
-
-            #align flipped calf to flipped middle leg to move the head
-            flipedcalf.align_orientation(flippeddigi1)
-
-            flipedcalf.length = flippeddigi1.length
-
-            #assign calf tail to flipped calf head so it moves calf's head
-            calf.head = flipedcalf.tail
-
-            #delete helper bones
-            bpy.ops.armature.select_all(action='DESELECT')
-            flippeddigi1.select = True
-            bpy.ops.armature.delete()
-            bpy.ops.armature.select_all(action='DESELECT')
-            flipedcalf.select = True
-            bpy.ops.armature.delete()
-            bpy.ops.armature.select_all(action='DESELECT')
-
-
-            #Tada! It's done! Now to duplicate toes and change parents
-
-
-            #duplicate old foot and reparent
-            newfoot = core.duplicatebone(digi3)
-            newfoot.parent = calf
-            #create toe bone if it exists
-            if(digi4):
-                #duplicate old toe and reparent
-                newtoe = core.duplicatebone(digi4)
-                newtoe.parent = newfoot
-            #finally done!
-        return {'FINISHED'}
-    
-    
-
-
-
-@wrapper_registry
-class Tuxedo_OT_DeleteZeroWeightBones(Operator):
-    bl_idname = 'tuxedo.delete_zero_weight_bones'
-    bl_label = t('Tools.delete_zero_weight_bones.label')
-    bl_description = t('Tools.delete_zero_weight_bones.desc')
-    bl_options = {'REGISTER', 'UNDO'}
-
-
-    @classmethod
-    def poll(self, context: bpy.types.Context):
-        return context.active_object and context.active_object.type == "ARMATURE" and (context.object.mode == "EDIT" or context.object.mode == "POSE" or context.object.mode == "OBJECT") and (len(context.active_object.children) > 0)
-
-    def execute(self, context: bpy.types.Context):
-        armature = context.active_object
-        data: bpy.types.Armature = armature.data
-        weighted_groups: dict[bpy.types.Object, set[str]] = core.get_zero_and_weight_vertex_groups(armature=armature, invert=True)
-
-        #find similar bones, where the bones are weighted to entirely nothing.
-        #https://stackoverflow.com/a/10066661
-        result: set[str] = set([i.name for i in data.bones])
-
-        for obj,groups in weighted_groups.items():
-            print([i for i in groups])
-            result = result - groups
-        
-        num: int = 0
-        prevmode = context.object.mode
-        core.Set_Mode(context,mode='EDIT')
-        for name in result:
-            data.edit_bones.remove(data.edit_bones[name])
-            num += 1
-        
-        core.Set_Mode(context,mode='OBJECT')
-        armature.update_tag(refresh={'DATA'})
-        core.Set_Mode(context,mode=prevmode)
-
-        self.report({'INFO'}, t('Tools.delete_zero_weight_bones.number').format(amount=num))
-
-        return {'FINISHED'}
-
-
-@wrapper_registry
-class Tuxedo_OT_DeleteZeroWeightVertexGroups(Operator):
-    bl_idname = 'tuxedo.delete_zero_weight_vertex_groups'
-    bl_label = t('Tools.delete_zero_weight_vertex_groups.label')
-    bl_description = t('Tools.delete_zero_weight_vertex_groups.desc')
-    bl_options = {'REGISTER', 'UNDO'}
-
-    @classmethod
-    def poll(self, context: bpy.types.Context):
-        return context.active_object and context.active_object.type == "ARMATURE" and (context.object.mode == "EDIT" or context.object.mode == "POSE" or context.object.mode == "OBJECT") and (len(context.active_object.children) > 0)
-
-    def execute(self, context: bpy.types.Context):
-        armature = context.active_object
-        zero_weight_vertex_groups: dict[bpy.types.Object, set[str]] = core.get_zero_and_weight_vertex_groups(armature=armature)
-        
-        num: int = 0
-
-        for obj,groups in zero_weight_vertex_groups.items():
-            for group in groups:
-                obj.vertex_groups.remove(obj.vertex_groups[group])
-                num += 1
-
-        self.report({'INFO'}, t('Tools.delete_zero_weight_vertex_groups.number').format(amount=num))
-        return {'FINISHED'}
-
-@wrapper_registry
-class Tuxedo_OT_ShapekeyToBasis(Operator):
-    bl_idname = 'tuxedo.shapekey_to_basis'
-    bl_label = t('Translations.shapekey_to_basis.label')
-    bl_description = t('Translations.shapekey_to_basis.desc')
-    bl_options = {'REGISTER', 'UNDO'}
-
-
-    @classmethod
-    def poll(self, context: bpy.types.Context):
-        return core.has_shapekeys(context.active_object) and hasattr(context.active_object.active_shape_key,'name')
-        
-    def execute(self, context: bpy.types.Context):
-        obj: bpy.types.Object = context.active_object
-        data: bpy.types.Mesh = obj.data
-        name: str = data.shape_keys.key_blocks[obj.active_shape_key_index].name
-        if not core.apply_shapekey_to_basis(context,obj,name,delete_old=False):
-            self.report({'ERROR'}, t('Translations.shapekey_to_basis.err1'))
-            return {'CANCELLED'}
-        else:
-            return {'FINISHED'}
-
-
-
-@wrapper_registry
-class Tuxedo_OT_TranslateMMD(Operator):
-    bl_idname = 'tuxedo.translate_mmd'
-    bl_label = t('Translations.translate_mmd.label')
-    bl_description = t('Translations.translate_mmd.desc')
-    bl_options = {'REGISTER', 'UNDO'}
-
-    translation_enum: bpy.props.EnumProperty(name=t('Translations.translation_coll_target.label'),description=t('Translations.translation_coll_target.desc'),items=translation_enum)
-    
-    def draw(self, context: bpy.types.Context):
-        self.layout.prop(self, "translation_enum")
-
-    def invoke(self, context, event):
-        return context.window_manager.invoke_props_dialog(self)
-
-    @classmethod
-    def poll(self, context: bpy.types.Context):
-        return (translation_enum(self, context) is not None) and (globals.mmd_tools_exist) and core.get_meshes_objects(context)
-
-    def execute(self, context: bpy.types.Context):
-        translate_mmd(self.translation_enum, context)
-        return {'FINISHED'}
-
-
-@wrapper_registry
-class Tuxedo_OT_ConvertToResonite(Operator):
-    bl_idname = 'tuxedo.convert_to_resonite'
-    bl_label = t('Tools.convert_to_resonite.label')
-    bl_description = t('Tools.convert_to_resonite.desc')
-    bl_options = {'REGISTER', 'UNDO'}
-
-    @classmethod
-    def poll(cls, context: bpy.types.Context):
-        if not core.get_armature(context):
-            return False
-        return True
-        
-    def execute(self, context: bpy.types.Context):
-        armature = core.get_armature(context)
-
-        translate_bone_fails = 0
-        untranslated_bones = set()
-
-        reverse_bone_lookup = dict()
-        for (preferred_name, name_list) in bone_names.items():
-            for name in name_list:
-                reverse_bone_lookup[name] = preferred_name
-
-        resonite_translations = {
-            'hips': "Hips",
-            'spine': "Spine",
-            'chest': "Chest",
-            'neck': "Neck",
-            'head': "Head",
-            'left_eye': "Eye.L",
-            'right_eye': "Eye.R",
-            'right_leg': "UpperLeg.R",
-            'right_knee': "Calf.R",
-            'right_ankle': "Foot.R",
-            'right_toe': 'Toes.R',
-            'right_shoulder': "Shoulder.R",
-            'right_arm': "UpperArm.R",
-            'right_elbow': "ForeArm.R",
-            'right_wrist': "Hand.R",
-            'left_leg': "UpperLeg.L",
-            'left_knee': "Calf.L",
-            'left_ankle': "Foot.L",
-            'left_toe': "Toes.L",
-            'left_shoulder': "Shoulder.L",
-            'left_arm': "UpperArm.L",
-            'left_elbow': "ForeArm.L",
-            'left_wrist': "Hand.R",
-
-            'pinkie_1_l': "pinkie1.L",
-            'pinkie_2_l': "pinkie2.L",
-            'pinkie_3_l': "pinkie3.L",
-            'ring_1_l': "ring1.L",
-            'ring_2_l': "ring2.L",
-            'ring_3_l': "ring3.L",
-            'middle_1_l': "middle1.L",
-            'middle_2_l': "middle2.L",
-            'middle_3_l': "middle3.L",
-            'index_1_l': "index1.L",
-            'index_2_l': "index2.L",
-            'index_3_l': "index3.L",
-            'thumb_1_l': "thumb1.L",
-            'thumb_2_l': "thumb2.L",
-            'thumb_3_l': "thumb3.L",
-
-            'pinkie_1_r': "pinkie1.R",
-            'pinkie_2_r': "pinkie2.R",
-            'pinkie_3_r': "pinkie3.R",
-            'ring_1_r': "ring1.R",
-            'ring_2_r': "ring2.R",
-            'ring_3_r': "ring3.R",
-            'middle_1_r': "middle1.R",
-            'middle_2_r': "middle2.R",
-            'middle_3_r': "middle3.R",
-            'index_1_r': "index1.R",
-            'index_2_r': "index2.R",
-            'index_3_r': "index3.R",
-            'thumb_1_r': "thumb1.R",
-            'thumb_2_r': "thumb2.R",
-            'thumb_3_r': "thumb3.R"
-        }
-
-                
-
-        context.view_layer.objects.active = armature
-        bpy.ops.object.mode_set(mode='EDIT')
-
-        bpy.ops.object.mode_set(mode='OBJECT')
-        for bone in armature.data.bones:
-            if core.simplify_bonename(bone.name) in reverse_bone_lookup and reverse_bone_lookup[core.simplify_bonename(bone.name)] in resonite_translations:
-                bone.name = resonite_translations[reverse_bone_lookup[core.simplify_bonename(bone.name)]]
-            else:
-                untranslated_bones.add(bone.name)
-                translate_bone_fails += 1
-            
-        bpy.ops.object.mode_set(mode='OBJECT')
-
-        if translate_bone_fails > 0:
-            self.report({'INFO'}, t('Tools.convert_bones_resonite.fail').format(translate_bone_fails=translate_bone_fails))
-        else:
-            self.report({'INFO'}, t('Tools.convert_bones.success'))
-
-
-        return {'FINISHED'}
-
-@wrapper_registry
-class Tuxedo_OT_MergeBoneWeightsToActive(Operator):
-    bl_idname = 'tuxedo.merge_weights_to_active'
-    bl_label = t('Tools.merge_weights_to_active.label')
-    bl_description = t('Tools.merge_weights_to_active.desc')
-    bl_options = {'REGISTER', 'UNDO'}
-
-    @classmethod
-    def poll(cls, context: bpy.types.Context):
-        return (bpy.context.mode == 'EDIT_ARMATURE') and (core.get_armature(context) is not None)
-
-    def execute(self, context: bpy.types.Context):
-        
-        armature: bpy.types.Armature = core.get_armature(context).data
-
-        original_edit = armature.use_mirror_x
-
-
-        core.Set_Mode(context, 'EDIT')
-        armature.use_mirror_x = False
-        core.merge_bone_weights(context, core.get_armature(context), bone_names=[i.name for i in context.selected_editable_bones], active_bone_name=context.active_bone.name, remove_old=context.scene.delete_old_bones_merging)
-
-        armature.use_mirror_x = original_edit
-        return {'FINISHED'}
-
-@wrapper_registry
-class Tuxedo_OT_MergeBoneWeightsToParents(Operator):
-    bl_idname = 'tuxedo.merge_weights_to_parents'
-    bl_label = t('Tools.merge_weights_to_parents.label')
-    bl_description = t('Tools.merge_weights_to_parents.desc')
-    bl_options = {'REGISTER', 'UNDO'}
-
-    @classmethod
-    def poll(cls, context: bpy.types.Context):
-        return (bpy.context.mode == 'EDIT_ARMATURE') and (core.get_armature(context) is not None)
-
-    def execute(self, context: bpy.types.Context):
-        armature: bpy.types.Armature = core.get_armature(context).data
-
-        original_edit = armature.use_mirror_x
-
-
-        core.Set_Mode(context, 'EDIT')
-        armature.use_mirror_x = False
-        core.merge_bone_weights_to_respective_parents(context, core.get_armature(context), bone_names=[i.name for i in context.selected_editable_bones], remove_old=context.scene.delete_old_bones_merging)
-        
-        armature.use_mirror_x = original_edit
-        return {'FINISHED'}
-
-#a vastly improved and simplified apply modifier for object with shape keys.
-@wrapper_registry
-class Tuxedo_OT_ApplyModifierForObjectWithShapeKeys(bpy.types.Operator):
-    bl_idname = 'tuxedo.apply_modifier_for_object_with_shape_keys'
-    bl_label = t('Tools.apply_modifier_for_object_with_shape_keys.label')
-    bl_description = t('Tools.apply_modifier_for_object_with_shape_keys.desc')
-    bl_options = {'REGISTER', 'UNDO'}
-
-    @classmethod
-    def poll(cls, context: bpy.types.Context):
-        return (bpy.context.mode == 'OBJECT') and (context.active_object is not None) and core.has_shapekeys(context.active_object)
-
-    modifiers: bpy.props.EnumProperty(name=t('Tools.modifiers_enum_label.label'), description=t('Tools.modifiers_enum_label.desc'), items=core.get_modifiers_active)
-
-    def draw(self, context):
-        self.layout.label(text=t('Tools.apply_modifier_for_object_with_shape_keys.drawlabel1').format(name=context.object.name))
-        self.layout.prop(self, "modifiers")
-
-    def execute(self, context: bpy.types.Context):
-        
-        result = core.apply_modifier_for_obj_with_shapekeys(context.object.modifiers[self.modifiers],delete_old=True)
-        if (result != True):
-            self.report({'ERROR'}, result)
-
-        return {'FINISHED'}
-    
-    def invoke(self, context, event):
-        return context.window_manager.invoke_props_dialog(self)
-
-@wrapper_registry
-class Tuxedo_OT_StartPoseMode(bpy.types.Operator):
-    bl_idname = 'tuxedo.start_pose_mode'
-    bl_label = t('Tools.start_pose_mode.label')
-    bl_description = t('Tools.start_pose_mode.desc')
-    bl_options = {'REGISTER', 'UNDO'}
-
-    @classmethod
-    def poll(cls, context: bpy.types.Context):
-        return core.get_armature(context)
-
-    def execute(self, context: bpy.types.Context):
-        armature = core.get_armature(context)
-
-        core.Set_Mode(context,'OBJECT')
-        core.unselect_all()
-        core.select(armature)
-        core.set_active(armature)
-        core.Set_Mode(context,'POSE')
-        #the next 3 lines allow us to clear the pose without ruining the user's selection.
-        bpy.ops.pose.transforms_clear()
-        core.select_set_all_curmode(context, 'INVERT')
-        bpy.ops.pose.transforms_clear()
-        return {'FINISHED'}
-
-@wrapper_registry
-class Tuxedo_OT_EndPoseMode(bpy.types.Operator):
-    bl_idname = 'tuxedo.end_pose_mode'
-    bl_label = t('Tools.end_pose_mode.label')
-    bl_description = t('Tools.end_pose_mode.desc')
-    bl_options = {'REGISTER', 'UNDO'}
-
-    @classmethod
-    def poll(cls, context: Context):
-        return core.get_armature(context)
-
-    def execute(self, context: Context):
-        bpy.ops.tuxedo.start_pose_mode() #take advantage of start pose method
-        core.Set_Mode(context,'OBJECT')
-        return {'FINISHED'}
-
-#didn't know this existed already as PoseToRest - @989onan
-"""@wrapper_registry
-class Tuxedo_OT_ApplyAsRest(bpy.types.Operator):
-    bl_idname = 'tuxedo.armature_apply_as_rest'
-    bl_label = t('Tools.armature_apply_as_rest.label')
-    bl_description = t('Tools.armature_apply_as_rest.desc')
-    bl_options = {'REGISTER', 'UNDO'}
-
-    @classmethod
-     def poll(cls, context: Context):
-        return core.get_armature(context)
-
-    def execute(self, context: Context):
-        meshes_and_armature = core.get_meshes_objects(context,core.get_armature(context).name )
-        armature = core.get_armature(context)
-        
-
-        for obj in meshes_and_armature:
-            if obj.type == 'MESH':
-                modifier = None
-                for mod in obj.modifiers:
-                    if mod.type == "ARMATURE" and mod.object == armature:
-                        modifier = mod
-                        break
-                
-                if modifier == None: continue
-                if core.has_shapekeys(obj):
-                    result = core.apply_modifier_for_obj_with_shapekeys(modifier)
-                    if (result != True):
-                        self.report({'ERROR'},result)
-                else:
-                    core.apply_modifier(modifier)
-        
-        core.Set_Mode(context,'OBJECT')
-        core.unselect_all()
-        core.select(armature)
-        core.set_active(armature)
-        core.Set_Mode(context,'POSE')
-
-        bpy.ops.pose.armature_apply()
-        core.Set_Mode(context,'OBJECT')
-        return {'FINISHED'} """
-
-@wrapper_registry
-class Tuxedo_OT_SavePoseAsShapekey(bpy.types.Operator):
-    bl_idname = 'tuxedo.save_pose_as_shapekey'
-    bl_label = t('Tools.save_pose_as_shapekey.label')
-    bl_description = t('Tools.save_pose_as_shapekey.desc')
-    bl_options = {'REGISTER', 'UNDO'}
-
-    @classmethod
-    def poll(cls, context):
-        return core.get_armature(context)
-
-    def execute(self, context):
-        meshes_and_armature = core.get_meshes_objects(context,core.get_armature(context).name )
-        armature = core.get_armature(context)
-        
-        core.Set_Mode(context,'OBJECT')
-        for obj in meshes_and_armature:
-            if obj.type == 'MESH':
-                
-                core.unselect_all()
-                core.select(obj)
-                core.set_active(obj)
-                modifier = None
-                for mod in obj.modifiers:
-                    if mod.type == "ARMATURE" and mod.object == armature:
-                        modifier = mod
-                        break
-                
-                if modifier == None: continue
-                bpy.ops.object.modifier_apply_as_shapekey(keep_modifier=True, modifier=modifier.name)
-
-        
-        core.unselect_all()
-        core.select(armature)
-        core.set_active(armature)
-        core.Set_Mode(context,'POSE')
-        return {'FINISHED'}
 
 @wrapper_registry
 class FitClothes(bpy.types.Operator):
