+++ ui.py
@@ -171,44 +171,6 @@ class ToolPanel(Panel):
         box = layout.box()
         col = box.column(align=True)
 
-        
-        row = col.row(align=True)
-        row.operator(tools.Tuxedo_OT_ApplyModifierForObjectWithShapeKeys.bl_idname)
-
-        row = col.row(align=True)
-        row.label(text=t('Tools.merge_bones.label'))
-        row.operator(tools.Tuxedo_OT_MergeBoneWeightsToActive.bl_idname)
-        row.operator(tools.Tuxedo_OT_MergeBoneWeightsToParents.bl_idname)
-        row = col.row(align=True)
-        row.prop(context.scene, 'delete_old_bones_merging')
-        
-        row = col.row(align=True)
-        row.label(text=t('Tools.translate.label'),icon="WORLD")
-        row = col.row(align=True)
-        row.operator(tools.ConvertToSecondlifeButton.bl_idname)
-        row.operator(gmod_tools.ConvertToValveButton.bl_idname)
-        row.operator(tools.Tuxedo_OT_ConvertToResonite.bl_idname)
-        row = col.row(align=True)
-        row.operator(tools.Tuxedo_OT_TranslateMMD.bl_idname)
-
-        row = col.row(align=False)
-        row.label(text=t('Tools.delete.label'),icon="TRASH")
-        row = col.row(align=False)
-        row.operator(tools.Tuxedo_OT_DeleteZeroWeightBones.bl_idname)
-        row = col.row(align=False)
-        row.operator(tools.Tuxedo_OT_DeleteZeroWeightVertexGroups.bl_idname)
-        row = col.row(align=False)
-        row.label(text=t('Tools.general.label'))
-        row = col.row(align=True)
-        row.operator(tools.Tuxedo_OT_ConnectBones.bl_idname)
-        row.operator(tools.Tuxedo_OT_DuplicateBones.bl_idname)
-        row = col.row(align=True)
-        row.operator(tools.Tuxedo_OT_CreateDigitigradeLegs.bl_idname)
-        row = col.row(align=True)
-        row.operator(tools.Tuxedo_OT_RemoveDoublesSafely.bl_idname)
-
-        row = col.row(align=True)
-        row.label(text=t('ToolPanel.decimation_panel.label'),icon='MOD_DECIM')
         row = col.row(align=True)
         row.prop(context.scene, 'decimation_animation_weighting', expand=True)
         if context.scene.decimation_animation_weighting:
